# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Education Center CRM API
  version: 1.0.0
  description: REST API for the MVP CRM platform of an education center.
servers:
  - url: http://localhost:4000/api
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPayload"
      responses:
        "201":
          description: User created
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPayload"
      responses:
        "200":
          description: User and JWT pair
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshPayload"
      responses:
        "200":
          description: New access and refresh tokens
...
  /children:
    get:
      tags: [Children]
      summary: List children
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of children
    post:
      tags: [Children]
      summary: Create a child
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChildPayload"
      responses:
        "201":
          description: Child created
  /children/{id}:
    patch:
      tags: [Children]
      summary: Update a child
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Child updated
    delete:
      tags: [Children]
      summary: Delete a child
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Child removed
...
  /enrollments:
    get:
      tags: [Enrollments]
      summary: List all enrollments for staff
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Enrollments list
    post:
      tags: [Enrollments]
      summary: Enroll a child into a lesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentPayload"
      responses:
        "201":
          description: Enrollment created
  /enrollments/my:
    get:
      tags: [Enrollments]
      summary: List current parent enrollments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Parent enrollments
...
  /attendance:
    post:
      tags: [Attendance]
      summary: Mark lesson attendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendancePayload"
      responses:
        "201":
          description: Attendance marked
  /attendance/{lessonId}:
    get:
      tags: [Attendance]
      summary: Get attendance journal for a lesson
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attendance list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterPayload:
      type: object
      required: [fullName, email, phone, password]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
    LoginPayload:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshPayload:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    ChildPayload:
      type: object
      required: [fullName, birthDate]
      properties:
        parentId:
          type: string
        fullName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        medicalNotes:
          type: string
...
    EnrollmentPayload:
      type: object
      required: [childId, lessonId]
      properties:
        childId:
          type: string
        lessonId:
          type: string
...
    AttendancePayload:
      type: object
      required: [enrollmentId, status]
      properties:
        enrollmentId:
          type: string
        status:
          type: string
          enum: [PRESENT, ABSENT]
        comment:
          type: string